Abstract KML importers a little to have more in the parent class.
